Longest Lines + Interesting Version
The Ruby version is lovely but takes O(logN) to sort the strings by length.

The more interesting version is O(N) and maintains a running list of the longest
known lines. Each new line is compared against the shortest known line to see if
it is viable, and if so, it is inserted to the proper position. If we've
accumulated beyond the line limit, the shortest line in the list is then
dropped.