Refactor Message Decode
`Message.decode` was pretty dense. This is a bit more code, but `Stream`
provides a nice API that simplifies the top level.

The `Dictionary` class doesn't simplify anything, but it does get the key/value
generation out of `Message`, where it doesn't quite belong.

My only real gripe is that `Stream` awkwardly looks ahead one character, which
was necessary to be able to register `empty?` before a `nil` got returned. On
the other hand, it DRYs up the segment and character loops that existed in
`decode`.